requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
altair==5.4.1
annotated-types==0.7.0
annoy==1.17.3
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
audioread==3.0.1
av==13.0.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.16
botocore==1.35.16
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
cleanlab==2.6.6
cleanvision==0.3.6
click==8.1.7
cloudpickle==3.0.0
cohere==5.9.1
colorama==0.4.6
colorcet==3.1.0
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dask==2024.8.2
databases==0.9.0
dataclasses-json==0.6.7
datasets==2.21.0
datashader==0.16.3
debugpy==1.8.5
decorator==5.1.1
deepdiff==8.0.1
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
effdet==0.4.1
emoji==2.12.1
et-xmlfile==1.1.0
executing==2.1.0
extra-streamlit-components==0.1.71
fastapi==0.114.1
fastavro==1.9.7
filelock==3.16.0
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
graphviz==0.20.3
greenlet==3.1.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==0.16.3
httptools==0.6.1
httpx==0.23.3
httpx-sse==0.4.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
imagecodecs==2024.6.1
ImageHash==4.3.1
imageio==2.35.1
importlib-resources==5.7.1
importlib_metadata==8.4.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kubernetes==30.1.0
langchain==0.2.16
langchain-chroma==0.1.3
langchain-cohere==0.2.4
langchain-community==0.2.16
langchain-core==0.2.38
langchain-experimental==0.0.65
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.117
layoutparser==0.3.4
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.40.1
locket==1.0.0
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multipledispatch==1.0.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.6.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.2
openai==1.44.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orderly-set==5.2.2
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pacmap==0.7.3
pandas==2.2.2
param==2.1.1
parameterized==0.9.0
parso==0.8.4
partd==1.4.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pi_heif==0.18.0
pikepdf==9.2.1
pillow==10.4.0
platformdirs==4.3.2
plotly==5.24.0
pooch==1.8.2
portalocker==2.10.1
posthog==3.6.5
prettytable==3.11.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
py-machineid==0.6.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycocotools==2.0.8
pycparser==2.22
pyct==0.5.0
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pydeck==0.9.1
pygltflib==1.16.2
Pygments==2.18.0
pynndescent==0.5.13
pypandoc==1.13
pyparsing==3.1.4
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.1
PyWavelets==1.7.0
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
rank-bm25==0.2.2
rapidfuzz==3.9.7
rapidocr-onnxruntime==1.3.24
referencing==0.35.1
regex==2024.7.24
renumics-spotlight==1.6.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.0.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.38.5
streamlit==1.38.0
streamlit-option-menu==0.3.13
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tiktoken==0.7.0
timm==1.0.9
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
trimesh==4.4.9
typer==0.12.5
types-requests==2.32.0.20240907
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
unstructured==0.15.10
unstructured-client==0.25.8
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
urllib3==2.2.2
uvicorn==0.30.6
validators==0.34.0
watchdog==4.0.2
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0.1
widgetsnbextension==4.0.13
win32-setctime==1.1.0
winregistry==1.1.1
wordcloud==1.9.3
wrapt==1.16.0
xarray==2024.9.0
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.11.1
zipp==3.20.1